Switch to a new theme to replace the old child theme
While the old theme, PPB Magnus, mostly looked alright (on a big enough
screen), I decided it was time to write a new theme for PPB after many
frustrations with it, frustrations that seemed to require a lot of
effort and a deep dive into the theme code to fix. As I saw it, there
were a few main problems:

1. Image Focus: Magnus, the base theme on which the *PPB* Magnus
additions were hacked on, is a photoblogging template. It just wasn't
designed for text-heavy (and image-light) sites like PPB. Text sizes and
margins always seemed too big and it was difficult to fit a reasonable
amount of content into a viewport. The theme styles got in the way of
fixes rather than being a useful starting point.

2. Responsiveness: There were several instances in which content was
mis-sized or cut off on small viewports, especially the bits that were
added later (the map, full-width images, etc).

3. Fragility: A lot of the things that were added to the theme later
weren't particularly reliable (up to and including `<h2>` centering!).
Even those that worked (for now) tended to be ad hoc or overcomplicated
to avoid going messing with the theme core. The child theme included.

So, I wrote a new theme with these considerations in mind. This is the
result. It doesn't do everything the old theme did, but it does
everything PPB needs,  and fits our use-case better. It's mobile-first,
responsive and has typographic decisions that make more sense for this
kind of site. There's still room to grow, but this seems like a solid
improvement for now.

[The name, 'Quotidian' is a play on the name 'Magnus': the theme
description proudly declared that 'Magnus' is Latin for 'grand'. I
settled on the exact opposite, a Latin (or Latin-derived?) word for
'everyday'.]